Share more data between PlayerInventory & Inventory
